[decontextualize] fix logic error of llm generation data, add later data augmentation steps and indexing
