+++ ctxcomp/models/__init__.py
@@ -1,0 +1,1 @@
+from .fidt5 import FiDT5
