+++ data_augmentation/utils.py
@@ -1,0 +1,64 @@
+import os
+import re
+import argparse
+import json
+from tqdm import tqdm
+from glob import glob
+
+def remove_citations(sent):
+    sent = sent.strip()
+    return re.sub(r"\[\d+", "", re.sub(r" \[\d+", "", sent)).replace(" |", "").replace("]", "")
+
+def normalize_texts(texts):
+    texts = remove_citations(texts)
+    texts = texts.strip()
+    pattern = re.compile(r"\n")
+    texts = re.sub(pattern, ' ', texts).strip()
+    pattern = re.compile(r"\s+")
+    texts = re.sub(pattern, ' ', texts).strip()
+    return texts
+
+def load_nuggets_and_claims(path):
+    data = json.load(open(path, 'r'))
+
+    nuggets = []
+    claims = []
+    for i, item in enumerate(data['data']):
+        example_id = item["example_id"]
+        if len(item['output']) == 0:
+            continue
+        outputs = item['output'][0].strip().split('\n')[:10]
+        ndocs = item['number_of_documents']
+        nuggets.append({
+            "example_id": example_id, 
+            "contents": [normalize_texts(n).strip() for n in outputs],
+            "type": 'nuggets',
+            "ndocs": ndocs
+        })
+        # "full_text": normalize_texts(item['full_text'][0]), # we dont need this
+
+        for j in range(1, ndocs+1): # the first one is nuggets
+            example_doc_id = f"{example_id}#{j}" 
+            outputs = item['output'][j].strip().split('\n')[:10]
+            claims.append({
+                "example_doc_id": example_doc_id, 
+                "full_text": normalize_texts(item['full_text'][j]),
+                "contents": [normalize_texts(n).strip() for n in outputs],
+                "type": 'claims',
+            })
+    return nuggets, claims
+
+def load_question(path):
+    data = json.load(open(path, 'r'))
+
+    questions = []
+    for i, item in enumerate(data['data']):
+        example_id = f"mds-question_{data['args']['shard']}-{i}" 
+        outputs = item['output'].strip().split('?')[0] + "?"
+        questions.append({
+            "id": example_id, 
+            "contents": outputs,
+            "type": 'question',
+        })
+
+    return questions
