--- recomp/models/__init__.py
@@ -1,1 +1,0 @@
-from .fidt5 import FiDT5
