+++ result/eli5-Meta-Llama-3-8B-Instruct-oracle-0shotx0-top5-text-42-quick_test30.json.score
@@ -1,0 +1,9 @@
+{
+    "length": 166.26666666666668,
+    "str_em": 0,
+    "str_hit": 0,
+    "rougeLsum": 21.812677921737055,
+    "citation_rec": 34.33826358826358,
+    "citation_prec": 43.098142598142594,
+    "claims_nli": 22.22222222222222
+}
