+++ result/eli5-Meta-Llama-3-8B-Instruct-oracle-1shotx5-top5-text-42-quick_test30.json.score
@@ -1,0 +1,9 @@
+{
+    "length": 133.5,
+    "str_em": 0,
+    "str_hit": 0,
+    "rougeLsum": 23.95995837017472,
+    "citation_rec": 42.269841269841265,
+    "citation_prec": 40.5575998075998,
+    "claims_nli": 21.11111111111111
+}
