+++ result/eli5-Meta-Llama-3-8B-Instruct-oracle-summ-0shotx0-top5-summary-42-quick_test30.json.score
@@ -1,9 +1,9 @@
 {
-    "length": 131.33333333333334,
+    "length": 127.86666666666666,
     "str_em": 0,
     "str_hit": 0,
-    "rougeLsum": 20.833741712521554,
-    "citation_rec": 27.635022385022385,
-    "citation_prec": 39.5818054494525,
-    "claims_nli": 17.777777777777775
+    "rougeLsum": 22.7965795341882,
+    "citation_rec": 34.74981962481964,
+    "citation_prec": 45.77922077922078,
+    "claims_nli": 21.11111111111111
 }
