+++ result/eli5-Meta-Llama-3-8B-Instruct-oracle-summ-1shotx5-top5-summary-42-quick_test30.json.score
@@ -1,9 +1,9 @@
 {
-    "length": 91.96666666666667,
+    "length": 103.0,
     "str_em": 0,
     "str_hit": 0,
-    "rougeLsum": 20.375318867390003,
-    "citation_rec": 45.95959595959596,
-    "citation_prec": 49.00793650793652,
-    "claims_nli": 18.888888888888893
+    "rougeLsum": 20.464469283421906,
+    "citation_rec": 39.10113035113035,
+    "citation_prec": 44.913179413179414,
+    "claims_nli": 22.22222222222222
 }
